agents/tools/assistant/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
dateparser>=1.2.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
